Floating inspection panel with integrated tags
Replace StickyInspectionBar with FloatingInspectionPanel: a fixed,
centered panel floating above the viewport bottom that merges object
tags and inspection controls into a single persistent toolbar.

- Lift useInspectionState from TargetTab to UnifiedObjectPage
- Auto-select singleton target on overview tab for inspection
- Grey out inspection controls on multi-target overview with
  "Select a target to inspect" message
- Bigger fonts (text-xs→text-sm, text-sm→text-base), rounded-xl shadow
- Add dropdownPlacement='top' prop to FilterChip and ObjectListsSection
  so dropdowns open upward from the bottom-anchored panel
- Remove ObjectListsSection from page header (now in floating panel)
- Add pb-24 content padding to avoid panel overlap
- Delete StickyInspectionBar (fully replaced)
